# swagger.json
{
    "swagger":"2.0",
    "info":{
        "version":"1.0.0",
        "title":"csbn API",
        "description":"csbn API"
    },
    "basePath":"/",
    "tags":[
        {
            "name":"Tools",
            "description":"API for tools in the system"
        },
        {
            "name":"Product",
            "description":"API for Product in the system"
        },
        {
            "name":"Batch",
            "description":"API for Batch in the system"
        },
        {
            "name":"Invoice",
            "description":"API for invoice in the system"
        },
        {
            "name":"Shipment",
            "description":"API for shipment in the system"
        },
        {
            "name":"Slaughterer",
            "description":"API for slaughterer in the system"
        },
        {
            "name":"Image",
            "description":"API for Image in the system"
        }
    ],
    "schemes":[
        "http"
    ],
    "consumes":[
        "application/json"
    ],
    "produces":[
        "application/json"
    ],
    "paths":{
        "/reach/{channel}/{org}":{
            "parameters":[
                {
                    "$ref":"#/parameters/channel"
                },
                {
                    "$ref":"#/parameters/organization"
                }
            ],
            "get":{
                "tags":[
                    "Tools"
                ],
                "description":null,
                "parameters":[

                ],
                "responses":{

                }
            }
        },
        "/createProduct/{channel}/{org}":{
            "parameters":[
                {
                    "$ref":"#/parameters/channel"
                },
                {
                    "$ref":"#/parameters/organization"
                }
            ],
            "post":{
                "tags":[
                    "Product"
                ],
                "description": null,
                "parameters":[
                    {
                        "name": "product",
                        "in": "body",
                        "description": "Product that we want to create",
                        "schema":{
                            "$ref":"#/definitions/Product"
                        }
                    }
                ],
                "responses":{

                }
            }
        },
        "/readProduct/{channel}/{org}/{id}":{
            "parameters":[
                {
                    "$ref":"#/parameters/channel"
                },
                {
                    "$ref":"#/parameters/organization"
                },
                {
                    "name":"id",
                    "in":"path",
                    "required":true,
                    "description":"ID of Products that we want to find",
                    "type":"string"
                }
            ],
            "get":{
                "tags":[
                    "Product"
                ],
                "description":null,
                "responses":{

                }
            }
        },
        "/readProducts/{channel}/{org}":{
            "parameters":[
                {
                    "$ref":"#/parameters/channel"
                },
                {
                    "$ref":"#/parameters/organization"
                }
            ],
            "get":{
                "tags":[
                    "Product"
                ],
                "description":null,
                "responses":{

                }
            }
        },
        "/updateProduct/{channel}/{org}/{id}":{
            "parameters":[
                {
                    "$ref":"#/parameters/channel"
                },
                {
                    "$ref":"#/parameters/organization"
                },
                {
                    "name":"id",
                    "in":"path",
                    "required":true,
                    "description":"ID of Product that we want to update",
                    "type":"string"
                }
            ],
            "put":{
                "summary":"Update product with give ID",
                "tags":[
                    "Product"
                ],
                "parameters":[
                    {
                        "name":"formData",
                        "in":"body",
                        "description":"product that we want to create",
                        "schema":{
                            "$ref":"#/definitions/Product"
                        }
                    }
                ],
                "responses":{

                }
            }
        },

        "/createBatch/{channel}/{org}":{
            "parameters":[
                {
                    "$ref":"#/parameters/channel"
                },
                {
                    "$ref":"#/parameters/organization"
                }
            ],
            "post":{
                "tags":[
                    "Batch"
                ],
                "description":null,
                "parameters":[
                    {
                        "name":"batch",
                        "in":"body",
                        "description":"Batch that we want to create",
                        "schema":{
                            "$ref":"#/definitions/Batch"
                        }
                    }
                ],
                "responses":{

                }
            }
        },
        "/readBatch/{channel}/{org}/{id}":{
            "parameters":[
                {
                    "$ref":"#/parameters/channel"
                },
                {
                    "$ref":"#/parameters/organization"
                },
                {
                    "name":"id",
                    "in":"path",
                    "required":true,
                    "description":"ID of Batch that we want to find",
                    "type":"string"
                }
            ],
            "get":{
                "tags":[
                    "Batch"
                ],
                "description":null,
                "responses":{

                }
            }
        },
		"/readBatches/{channel}/{org}":{
            "parameters":[
                {
                    "$ref":"#/parameters/channel"
                },
                {
                    "$ref":"#/parameters/organization"
                }
            ],
            "get":{
                "tags": [
                    "Batch"
                ],
                "description": null,
                "responses":{

                },
				"parameters": [
                    {
                        "name": "date",
                        "in": "query",
                        "description": "date range",
						"type": "string"
                    }
                ]
            }
        },
        "/updateBatch/{channel}/{org}/{id}":{
            "parameters":[
                {
                    "$ref": "#/parameters/channel"
                },
                {
                    "$ref": "#/parameters/organization"
                },
                {
                    "name":"id",
                    "in":"path",
                    "required":true,
                    "description":"ID of Batch that we want to update",
                    "type":"string"
                }
            ],
            "put":{
                "summary":"Update batch with give ID",
                "tags":[
                    "Batch"
                ],
                "parameters":[
                    {
                        "name":"formData",
                        "in":"body",
                        "description":"batch that we want to create",
                        "schema":{
                            "$ref":"#/definitions/Batch"
                        }
                    }
                ],
                "responses":{

                }
            }
        },

        "/createInvoice/{channel}/{org}":{
            "parameters":[
                {
                    "$ref":"#/parameters/channel"
                },
                {
                    "$ref":"#/parameters/organization"
                }
            ],
            "post":{
                "tags":[
                    "Invoice"
                ],
                "description":null,
                "parameters":[
                    {
                        "in":"body",
                        "description":"Invoice that we want to create",
                        "schema":{
                            "$ref":"#/definitions/Invoice"
                        }
                    }
                ],
                "responses":{

                }
            }
        },
        "/readInvoice/{channel}/{org}/{id}":{
            "parameters":[
                {
                    "$ref":"#/parameters/channel"
                },
                {
                    "$ref":"#/parameters/organization"
                },
                {
                    "name":"id",
                    "in":"path",
                    "required":true,
                    "description":"ID of Invoice that we want to find",
                    "type":"string"
                }
            ],
            "get":{
                "tags":[
                    "Invoice"
                ],
                "description":null,
                "responses":{

                }
            }
        },
        "/updateInvoice/{channel}/{org}/{id}":{
            "parameters":[
                {
                    "$ref":"#/parameters/channel"
                },
                {
                    "$ref":"#/parameters/organization"
                },
                {
                    "name":"id",
                    "in":"path",
                    "required":true,
                    "description":"ID of Invoice that we want to update",
                    "type":"string"
                }
            ],
            "put":{
                "summary":"Update Invoice with give ID",
                "tags":[
                    "Invoice"
                ],
                "parameters":[
                    {
                        "in":"body",
                        "description":"invoice that we want to create",
                        "schema":{
                            "$ref":"#/definitions/Invoice"
                        }
                    }
                ],
                "responses":{

                }
            }
        },

        "/createShipment/{channel}/{org}":{
            "parameters":[
                {
                    "$ref":"#/parameters/channel"
                },
                {
                    "$ref":"#/parameters/organization"
                }
            ],
            "post":{
                "tags":[
                    "Shipment"
                ],
                "description":null,
                "parameters":[
                    {
                        "name":"shipment",
                        "in":"body",
                        "description":"Shipment that we want to create",
                        "schema":{
                            "$ref":"#/definitions/Shipment"
                        }
                    }
                ],
                "responses":{

                }
            }
        },
        "/readShipment/{channel}/{org}/{id}":{
            "parameters":[
                {
                    "$ref":"#/parameters/channel"
                },
                {
                    "$ref":"#/parameters/organization"
                },
                {
                    "name":"id",
                    "in":"path",
                    "required":true,
                    "description":"ID of shipment that we want to find",
                    "type":"string"
                }
            ],
            "get":{
                "tags":[
                    "Shipment"
                ],
                "description":null,
                "responses":{

                }
            }
        },
        "/updateShipment/{channel}/{org}/{id}":{
            "parameters":[
                {
                    "$ref":"#/parameters/channel"
                },
                {
                    "$ref":"#/parameters/organization"
                },
                {
                    "name":"id",
                    "in":"path",
                    "required":true,
                    "description":"ID of Shipment that we want to update",
                    "type":"string"
                }
            ],
            "put":{
                "summary":"Update shipment with give ID",
                "tags":[
                    "Shipment"
                ],
                "parameters":[
                    {
                        "name":"formData",
                        "in":"body",
                        "description":"Shipment that we want to update",
                        "schema":{
                            "$ref":"#/definitions/Shipment"
                        }
                    }
                ],
                "responses":{

                }
            }
        },
		"/createSlaughterer/{channel}/{org}":{
            "parameters":[
                {
                    "$ref":"#/parameters/channel"
                },
                {
                    "$ref":"#/parameters/organization"
                }
            ],
            "post":{
                "tags":[
                    "Slaughterer"
                ],
                "description":null,
                "parameters":[
                    {
                        "name":"slaughterer",
                        "in":"body",
                        "description":"Slaughterer that we want to create",
                        "schema":{
                            "$ref":"#/definitions/Slaughterer"
                        }
                    }
                ],
                "responses":{

                }
            }
        },
        "/readSlaughterer/{channel}/{org}/{id}":{
            "parameters":[
                {
                    "$ref":"#/parameters/channel"
                },
                {
                    "$ref":"#/parameters/organization"
                },
                {
                    "name":"id",
                    "in":"path",
                    "required":true,
                    "description":"ID of Slaughterer that we want to find",
                    "type":"string"
                }
            ],
            "get":{
                "tags":[
                    "Slaughterer"
                ],
                "description":null,
                "responses":{

                }
            }
        },
		"/readSlaughterers/{channel}/{org}":{
            "parameters":[
                {
                    "$ref":"#/parameters/channel"
                },
                {
                    "$ref":"#/parameters/organization"
                }
            ],
            "get":{
                "tags":[
                    "Slaughterer"
                ],
                "description":null,
                "responses":{

                }
            }
        },
        "/updateSlaughterer/{channel}/{org}/{id}":{
            "parameters":[
                {
                    "$ref":"#/parameters/channel"
                },
                {
                    "$ref":"#/parameters/organization"
                },
                {
                    "name":"id",
                    "in":"path",
                    "required":true,
                    "description":"ID of Slaughterer that we want to update",
                    "type":"string"
                }
            ],
            "put":{
                "summary":"Update Slaughterer with give ID",
                "tags":[
                    "Slaughterer"
                ],
                "parameters":[
                    {
                        "name":"formData",
                        "in":"body",
                        "description":"Slaughterer that we want to update",
                        "schema":{
                            "$ref":"#/definitions/Slaughterer"
                        }
                    }
                ],
                "responses":{

                }
            }
        },
		"/image/upload":{
            "post":{
                "tags":[
                    "Image"
                ],
                "description":null,
				"consumes": ["multipart/form-data"],
                "parameters":[
                    {
                        "name":"imgFile",
                        "in":"formData",
                        "description":"img file",
						"type": "file"
                    }
                ],
                "responses":{

                }
            }
        }
    },
    "parameters":{
        "channel":{
            "name":"channel",
            "in":"path",
            "required":true,
            "description":"Channel name",
            "type":"array",
            "items":{
                "type":"string",
                "enum":[
                    "channel0",
                    "channel1",
                    "channel2",
                    "channel3",
                    "channel4"
                ]
            }
        },
        "organization":{
            "name":"org",
            "in":"path",
            "required":true,
            "description":"Organization name",
            "type":"array",
            "items":{
                "type":"string",
                "enum":[
                    "courier",
                    "supp1",
                    "supp2",
                    "supp3",
                    "cust1",
                    "cust2",
                    "cust3"
                ]
            }
        }
    },
    "definitions":{
        "Product":{
            "required":[
                "id"
            ],
            "properties":{
                "id":{
                    "type":"string",
                    "uniqueItems":true
                },
                "name":{
                    "type":"string"
                },
				"imgPath":{
                    "type":"string"
                },
                "ingredients":{
                    "type":"array",
                    "items":{
                        "type":"string"
                    }
                },
                "halalCertificateId":{
                    "type":"string"
                }
            }
        },
        "Batch":{
            "required":[
                "id",
				"date"
            ],
            "properties":{
                "id":{
                    "type":"string",
                    "uniqueItems":true
                },
                "productState":{
                    "$ref":"#/definitions/Product"
                },
				"date":{
                    "type":"string"
                },
                "qty":{
                    "type":"integer"
                },
                "ingredients":{
                    "type":"array",
                    "items":{
                        "type":"string"
                    }
                },
                "invoices":{
                    "type":"array",
                    "items":{
                        "type":"object",
                        "properties":{
                            "id":{
                                "type":"string"
                            },
                            "qty":{
                                "type":"integer"
                            }
                        }
                    }
                }
            }
        },
        "Shipment":{
            "required":[
                "id",
				"date"
            ],
            "properties":{
                "id":{
                    "type":"string",
                    "uniqueItems":true
                },
				"date":{
                    "type":"string"
                },
                "invoices":{
                    "type":"array",
                    "items":{
                        "type":"object",
                        "properties":{
                            "id":{
                                "type":"string"
                            },
                            "items":{
                                "type":"array",
                                "items":{
                                    "$ref":"#/definitions/Item"
                                }
                            },
                            "isComplete":{
                                "type":"boolean"
                            }
                        }
                    }
                },
                "history":{
                    "type":"array",
                    "items":{
                        "type":"object",
                        "properties":{
                            "date":{
                                "type":"string"
                            },
                            "location":{
                                "type":"string"
                            }
                        }
                    }
                }
            }
        },
        "Invoice":{
            "required":[
                "id",
				"date"
            ],
            "properties":{
                "id":{
                    "type":"string",
                    "uniqueItems":true
                },
				"date":{
                    "type":"string"
                },
                "sellerId":{
                    "type":"string"
                },
                "buyerId":{
                    "type":"string"
                },
                "courierId":{
                    "type":"string"
                },
                "items":{
                    "type":"array",
                    "items":{
                        "$ref":"#/definitions/Item"
                    }
                },
                "shipmentState":{
                    "type":"array",
                    "items":{
                        "type":"object",
                        "properties":{
                            "id":{
                                "type":"string"
                            },
                            "items":{
                                "type":"array",
                                "items":{
                                    "$ref":"#/definitions/Item"
                                }
                            }
                        }
                    }
                }
            }
        },
        "Item":{
            "required":[
                "batchId"
            ],
            "properties":{
                "batchId":{
                    "type":"string"
                },
                "qty":{
                    "type":"integer"
                }
            }
        },
		"Slaughterer":{
            "required":[
                "id"
            ],
            "properties":{
                "id":{
                    "type":"string"
                },
				"name":{
                    "type":"string"
                },
                "imgUrl":{
                    "type":"string"
                },
                "certificateNo":{
                    "type":"string"
                },
				"certificateImgUrl":{
                    "type":"string"
                }
            }
        }
    }
}